Fix UTC offset sign error in sunrise/sunset calculation
time_t is unsigned, so subtracting the UTC time from local time wrapped
for zones west of Greenwich, producing a garbage offset (17236 instead of
-240 minutes). All sun events were computed 196 minutes late, so the
morning brightness ramp ran from ~8:20 to ~9:55 instead of ending at
sunrise. Cast the difference to int32_t before dividing.

+++ alphafive/examples/AlphaClock/AlphaClock.ino
@@ -1666,8 +1666,10 @@ void recomputeSunTimes(void)
   float lon = storedLon100 / 100.0;
   time_t tLocal = now();
 
-  // Current UTC offset (including DST), in minutes, from the active time zone
-  int utcOffsetMin = (int)((tLocal - timezones[tzIndex]->toUTC(tLocal)) / 60);
+  // Current UTC offset (including DST), in minutes, from the active time zone.
+  // time_t is unsigned, so the difference must be cast to a signed type before
+  // dividing: for zones west of Greenwich it is negative.
+  int utcOffsetMin = (int)(((int32_t)(tLocal - timezones[tzIndex]->toUTC(tLocal))) / 60);
 
   sunriseMinutes = sunEventMinutes(1, year(tLocal), month(tLocal), day(tLocal), lat, lon, utcOffsetMin, 90.833);
   sunsetMinutes = sunEventMinutes(0, year(tLocal), month(tLocal), day(tLocal), lat, lon, utcOffsetMin, 90.833);
